smooth estimation for duration of finishFrame()
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@3128 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/video/PixelRenderer.hh
@@ -212,7 +212,7 @@ private:
 	int nextY;
 
 	int frameSkipCounter;
-	unsigned finishFrameDuration;
+	float finishFrameDuration;
 	
 	static const unsigned NUM_FRAME_DURATIONS = 50;
 	CircularBuffer<unsigned, NUM_FRAME_DURATIONS> frameDurations;
